Scan fixes batch 1: the 7 code HIGHs + regression tests
- compMul kernel takes (target, from) like every sibling; mulInPlace(T,T)
  mutates the receiver; the four operator* call sites that compensated for
  the old swapped order flipped in the same change (fProxy/iProxy twins).
- Eigen.valuesQR -> valuesQRInPlace (destroys A; the InPlace suffix is the
  destruction signal). Cross-refs, messages, tests, benchmark updated.
- Blas.dotRows validates rows against a.M_Rows and c.M_Rows (OOB read+write).
- ILU0 Apply backward sweep: stackalloc hoisted out of the per-block-row loop
  (stack grew nb*16*sizeof per call).
- Scalar-shifted-by-vector bit shifts compute at the proxy's own width
  (long variant was masked mod 32 and truncated to int).
- iProxyLinVec interpolates in double (float lerp corrupted interior values).
- bool Analysis.isDiagonal now tests true diagonality (off-diagonal all-false
  + squareness); undocumented compare inverter removed.

New tests: mulInPlace/operator* value+aliasing regressions (InPlaceOpTests),
per-type exact-oracle scalar-shift tests incl. 1<<62 (CompBitsTests),
LinVec exactness with (1L<<40)+4 endpoint (InitTest). UIntTypeTests updated:
it had pinned the old swapped mulInPlace convention as intended behavior.

Suite: 6040/6040 green.

+++ Assets/LinearAlgebra/Source/Arena/ArenaExtensions.int.cs
@@ -73,9 +73,11 @@ namespace LinearAlgebra
             // Generate.linspace convention: a single sample returns {start}.
             if (N == 1) { vec[0] = start; return vec; }
 
-            float scale = 1 / (float)(N - 1);
+            // Interpolate in double: exact for every int/uint value and for long magnitudes
+            // up to 2^53 (float's 24-bit mantissa corrupts interior values).
+            double scale = 1.0 / (N - 1);
             for(int i = 0; i < N; i++) {
-                vec[i] = (int)math.lerp((int)start, (int)end, i * scale);
+                vec[i] = (int)math.lerp((double)start, (double)end, i * scale);
             }
             // Pin endpoints exactly (the lerp at the last index lands ~1 ulp short of end,
             // which can truncate to the wrong integer).
